Refactor schemata and add ref-based validation
